docs: cpp-tagfiles is antora extension
